DRAFT for Road-Rally OpenRally GPX
Proposal for a modern Road-Rally GPX standard. As far as I know, the old v0.3-DRAFT is not in active use, so I used the current v1.0.2 Cross-Country format as a source and modified it to fit the 2025 FIA Regional Rally Sporting Regulations. I added Waypoints for Chicane, Virtual Chicane (VC), and Slow Zone (SZ). This format differs from the Cross-Country format, so I added an openrally:type tag to differentiate between them.